Remove unwanted files and update remaining files
After forking the repo from the original source, I removed unwanted
files and updated the remaining files to improve the project's structure

I needed to troubleshoot an issue with the gemfile. Know that the forked
repo was not working properly initially. I diagnosed that the issue was
related to line 29 of the gemfile, which involved something named `wdm`

I resolved the issue by installing the latest version of `wdm` with this
command: `gem install wdm -v 0.2.0` and then updating line 29 to reflect
the new installation

Afterwards I made several changes to the files in the `_posts`,
`_drafts`, `_includes`, and `_pages` directories.